fix(04-06): clean static analysis gates (clang-tidy + cppcheck)
- Fix bugprone-incorrect-roundings: use lround() in config.cpp, keyhunt.cpp, output.cpp
- Fix bugprone-integer-division: explicit double casts in bloom_wrapper.h, bloom_init.cpp, mode_bsgs.cpp
- Fix bugprone-signed-char-misuse: cast to unsigned char in progress.cpp
- Fix null pointer dereference after calloc in address_util.cpp (cppcheck finding)
- Add default case to switch in io.cpp
- Add NOLINT for false positives: infinite-loop in search_minikeys.cpp, memcpy in sha256_shani.cpp
- Suppress noisy checks in .clang-tidy (macro-parentheses, narrowing-conversions, etc.)
- Update cppcheck Makefile target to exclude vendored code

+++ src/crypto/bloom_init.cpp
@@ -57,7 +57,7 @@ bool initBloomFilter(struct bloom *bloom_arg, uint64_t items_bloom) {
 			output_warning("--------------------------------------------------------\n");
 
 			/* Calculate optimal multiplier that fits */
-			int suggested_multiplier = (int)(safe_limit_mb * 1024 * 1024 / items_bloom / 3.59);
+			int suggested_multiplier = (int)((double)safe_limit_mb * 1024.0 * 1024.0 / (double)items_bloom / 3.59);
 			if (suggested_multiplier < 1) suggested_multiplier = 1;
 
 			output_warning("Try reducing -z parameter to: %d\n", suggested_multiplier);
